+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next version
+# 0.13.0
 + Move datatable to separate component
 + Add orderBy and orderWay to fetchAll
 + Move loading component. Refactor RolesPresentationals